cap quyen

+++ app/Http/Controllers/ChiTietCapQuyenController.php
@@ -9,6 +9,16 @@ class ChiTietCapQuyenController extends Controller
 {
     public function loadchiTietChucNang($id_chuc_vu)
     {
+        $id_chuc_nang = 1;
+        $user = $this->isUserAdmin();
+        $checkQuyen = ChiTietCapQuyen::where('id_chuc_vu', $user->id_chuc_vu)->where('id_chuc_nang', $id_chuc_nang)->first();
+        if (!$checkQuyen) {
+            return response()->json([
+                'message'  =>   'Bạn chưa được cấp quyền này',
+                'status'   =>   false,
+            ]);
+        }
+
         $data = ChiTietCapQuyen::where('id_chuc_vu', $id_chuc_vu)
                                 ->join('chuc_nangs', 'chuc_nangs.id', 'chi_tiet_cap_quyens.id_chuc_nang')
                                 ->select('chi_tiet_cap_quyens.*', 'chuc_nangs.ten_chuc_nang')
@@ -20,6 +30,15 @@ class ChiTietCapQuyenController extends Controller
     }
     public function store(Request $request)
     {
+        $id_chuc_nang = 1;
+        $user = $this->isUserAdmin();
+        $checkQuyen = ChiTietCapQuyen::where('id_chuc_vu', $user->id_chuc_vu)->where('id_chuc_nang', $id_chuc_nang)->first();
+        if (!$checkQuyen) {
+            return response()->json([
+                'message'  =>   'Bạn chưa được cấp quyền này',
+                'status'   =>   false,
+            ]);
+        }
 
         $check = ChiTietCapQuyen::where('id_chuc_vu', $request->id_chuc_vu)
                                  ->where('id_chuc_nang', $request->id_chuc_nang)
@@ -43,6 +62,16 @@ class ChiTietCapQuyenController extends Controller
     }
     public function destroy(Request $request)
     {
+        $id_chuc_nang = 1;
+        $user = $this->isUserAdmin();
+        $checkQuyen = ChiTietCapQuyen::where('id_chuc_vu', $user->id_chuc_vu)->where('id_chuc_nang', $id_chuc_nang)->first();
+        if (!$checkQuyen) {
+            return response()->json([
+                'message'  =>   'Bạn chưa được cấp quyền này',
+                'status'   =>   false,
+            ]);
+        }
+
         ChiTietCapQuyen::find($request->id)->delete();
 
         return response()->json([
