+++ app/Http/Controllers/ChucVuController.php
@@ -2,13 +2,25 @@
 
 namespace App\Http\Controllers;
 
+use App\Models\ChiTietCapQuyen;
 use App\Models\ChucVu;
+use Illuminate\Support\Facades\Auth;
 use Illuminate\Http\Request;
 
 class ChucVuController extends Controller
 {
     public function getDataChucVu()
     {
+        $id_chuc_nang = 1;
+        $user = $this->isUserAdmin();
+        $checkQuyen = ChiTietCapQuyen::where('id_chuc_vu', $user->id_chuc_vu)->where('id_chuc_nang', $id_chuc_nang)->first();
+        if (!$checkQuyen) {
+            return response()->json([
+                'message'  =>   'Bạn chưa được cấp quyền này',
+                'status'   =>   false,
+            ]);
+        }
+
         $data = ChucVu::select()->get();
         return response()->json([
             'data' => $data,
@@ -17,6 +29,16 @@ class ChucVuController extends Controller
 
     public function createDataChucVu(Request $request)
     {
+        $id_chuc_nang = 1;
+        $user = $this->isUserAdmin();
+        $checkQuyen = ChiTietCapQuyen::where('id_chuc_vu', $user->id_chuc_vu)->where('id_chuc_nang', $id_chuc_nang)->first();
+        if (!$checkQuyen) {
+            return response()->json([
+                'message'  =>   'Bạn chưa được cấp quyền này',
+                'status'   =>   false,
+            ]);
+        }
+
         ChucVu::create([
             'ten_chuc_vu' => $request->ten_chuc_vu,
         ]);
@@ -28,9 +50,18 @@ class ChucVuController extends Controller
 
     public function deleteChucVu($id)
     {
+        $id_chuc_nang = 1;
+        $user = $this->isUserAdmin();
+        $checkQuyen = ChiTietCapQuyen::where('id_chuc_vu', $user->id_chuc_vu)->where('id_chuc_nang', $id_chuc_nang)->first();
+        if (!$checkQuyen) {
+            return response()->json([
+                'message'  =>   'Bạn chưa được cấp quyền này',
+                'status'   =>   false,
+            ]);
+        }
         $ten_chuc_vu = ChucVu::where('id', $id)->first();
 
-        if($ten_chuc_vu) {
+        if ($ten_chuc_vu) {
             $ten_chuc_vu->delete();
 
             return response()->json([
@@ -46,8 +77,17 @@ class ChucVuController extends Controller
     }
     public function UpateChucVu(Request $request)
     {
+        $id_chuc_nang = 1;
+        $user = $this->isUserAdmin();
+        $checkQuyen = ChiTietCapQuyen::where('id_chuc_vu', $user->id_chuc_vu)->where('id_chuc_nang', $id_chuc_nang)->first();
+        if (!$checkQuyen) {
+            return response()->json([
+                'message'  =>   'Bạn chưa được cấp quyền này',
+                'status'   =>   false,
+            ]);
+        }
         $ten_chuc_vu = ChucVu::where('id', $request->id)->first();
-        if($ten_chuc_vu) {
+        if ($ten_chuc_vu) {
             $ten_chuc_vu->update([
                 'ten_chuc_vu'             => $request->ten_chuc_vu,
